# src/main/java/com/lcorp/shoppyservice/model/Role.java
package com.lcorp.shoppyservice.model;

import jakarta.persistence.*;

@Entity
@Table(name = "roles")
public class Role {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name="`id`")
    private Integer id;
    @Enumerated(EnumType.STRING)
    @Column(length = 20)
    private RoleEnum name;
    public Role() {
    }
    public Role(RoleEnum name) {
        this.name = name;
    }
    public Integer getId() {
        return id;
    }
    public void setId(Integer id) {
        this.id = id;
    }
    public RoleEnum getName() {
        return name;
    }
    public void setName(RoleEnum name) {
        this.name = name;
    }

    public Role(Integer id, RoleEnum name) {
        this.id = id;
        this.name = name;
    }
}